dsi-ws2025-project-grpab-weigl-mds1ab-awp-proj2-main/requirements.txt
ipython-sql
prettytable==2.5.0
psycopg2
psycopg2-binary
requests
python-dotenv
pandas
pymongo[srv]
numpy
streamlit
graphviz
bs4
matplotlib
seaborn
nltk
sentence_transformers
umap-learn
hdbscan
wordcloud
plotly
